Final Push -Ready for Review 2
Changed file names and added author and date to each file. Added JavaScript to run a timer that counts down one second at a time, and when it reaches the end of the timer, it displays HTML explaining that the order has expired. The JavaScript also makes sure that the numbers displayed in the timer have a zero before them if they are not double digits. The timer updates every second, and decreases the timer as such.